Refactor Server.php to remove unnecessary coolify.yaml from cloud hosted servers

+++ app/Models/Server.php
@@ -240,7 +240,7 @@ respond 404
         $dynamic_config_path = $this->proxyPath() . "/dynamic";
         if ($this->proxyType() === 'TRAEFIK_V2') {
             $file = "$dynamic_config_path/coolify.yaml";
-            if (empty($settings->fqdn) || isCloud()) {
+            if (empty($settings->fqdn) || (isCloud() && $this->id !== 0)) {
                 instant_remote_process([
                     "rm -f $file",
                 ], $this);
@@ -359,7 +359,7 @@ respond 404
             }
         } else if ($this->proxyType() === 'CADDY') {
             $file = "$dynamic_config_path/coolify.caddy";
-            if (empty($settings->fqdn)) {
+            if (empty($settings->fqdn) || (isCloud() && $this->id !== 0)) {
                 instant_remote_process([
                     "rm -f $file",
                 ], $this);
